refactored effects
